“The tenth saith unto him, Lord, how is it that thou wilt manifest thyself unto us, and not unto the world?”
